+++ src/refactorings/extract-function/index.ts
@@ -1,0 +1,11 @@
+import { RefactoringConfig } from "../../refactorings";
+import { extractFunction } from "./extract-function";
+
+const config: RefactoringConfig = {
+  command: {
+    key: "extractFunction",
+    operation: extractFunction
+  }
+};
+
+export default config;