+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- **[New Refactoring]** "Extract Function". It's merely a proxy to VS Code "Extract Function" refactoring. But it assigns a keybinding (`Ctrl + Alt + M` / `⌥ ⌘ M` on MacOS) for discoverability. It also transforms the code right away if there is a single possible extraction.
+- **[New Refactoring]** "Extract Function". It's merely a proxy to VS Code "Extract Function" refactoring. But it assigns a keybinding (`Ctrl + Alt + M` / `⌥ ⌘ M` on MacOS) for discoverability. It also transforms the code right away if there is a single possible extraction. Finally, it will expand the selection to the closest statement you can extract. So you don't have to _precisely_ select the statement you intend to extract.
 
 ## [9.5.0] - 2025-05-09 - Whistle While You Work 🛠️